+++ src/ast/passes/codegen_llvm.cpp
@@ -1310,8 +1310,7 @@ ScopedExpr CodegenLLVM::visit(Call &call)
     auto &map = *call.vargs.at(0).as<Map>();
     auto scoped_key = getMapKey(map, call.vargs.at(1));
 
-    Value *ret = b_.CreateMapDeleteElem(
-        map, scoped_key.value(), call.ret_val_discarded, call.loc);
+    Value *ret = b_.CreateMapDeleteElem(map, scoped_key.value());
     Value *expr = b_.CreateICmpEQ(ret, b_.getInt64(0), "delete_ret");
     return ScopedExpr(expr);
   } else if (call.func == "str") {